requirements.txt
python-magic
lxml
beautifulsoup4
argparse